reworks multigraph popup. adds offset for stacking.
Reworks the multigraph popup function to remove dependency on
an external variable. This removes a bug where the spinner would
not be removed if more than one popup was being opened before the
graph loaded. The count of the open graphs is being kept to help
with the offset of graphs for stacking -- however it is being
stuck on the global seldon object to avoid potential scope issues.