fixes bug in ie
jQuery focus event would cause issues in ie where the window would
be sent behind the other open windows. However this handler was
needed to prevent undesireable autofocus behavior in the jQuery UI
dialog windows. Fix is to removes calls to this handler and any
anchor tags possible in the dialog windows. If anchor tags are
needed (for example: the about page) then an autofocus attribute
must be put on some other element in that dialog's html.