Add support for overriding default styling for empty subgroups.
By setting break="true" on a wmsSubgroup in ews_config.xml,
you can now override the default behavior of seldon, which
adds an "empty" class to subgroups without any layers.

This accommodates an edge case where an empty subgroup is used
as a heading followed immediately by a collapsible subgroup.
By using break="true" in the subgroup config, the sublist will
add a buffer to the empty subgroup heading that will make
the layout look more consistent with the rest of the app.